requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
bidict==0.23.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
python-engineio==4.10.1
python-socketio==5.11.4
simple-websocket==1.1.0
typing_extensions==4.12.2
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.0
